Offloaded actions to test image. Separated some functionality. Removed unnecessary tests.
