+++ galasa-inttests-parent/dev.galasa.inttests/src/main/java/dev/galasa/inttests/compilation/AbstractCompilationLocal.java
@@ -11,16 +11,11 @@ import java.io.IOException;
 import java.net.URI;
 import java.net.URISyntaxException;
 import java.nio.charset.Charset;
-import java.nio.file.FileVisitResult;
 import java.nio.file.Files;
 import java.nio.file.Path;
-import java.nio.file.SimpleFileVisitor;
 import java.nio.file.StandardCopyOption;
-import java.nio.file.attribute.BasicFileAttributes;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
-import java.util.zip.ZipEntry;
-import java.util.zip.ZipInputStream;
 
 import org.apache.commons.logging.Log;
 import org.apache.http.ConnectionClosedException;
@@ -44,7 +39,7 @@ import dev.galasa.linux.LinuxManagerException;
 public abstract class AbstractCompilationLocal {
     
     @RunName
-    public String         runName;
+    public String         	runName;
     
     @Logger
     public Log              logger;
@@ -52,11 +47,8 @@ public abstract class AbstractCompilationLocal {
     @HttpClient
     public IHttpClient      client;
     
-    private String          prefix = "dev.galasa.simbank";
-    private String          testProjectName = prefix + ".tests";
-    private String          managerProjectName = prefix + ".manager";
-    
-    private Path            simplatformParentDir;
+    protected Path            testRunDirectory;
+    protected Path            projectDirectory;
     private Path            gradleBin;
     
     @TestProperty(prefix = "gradle.zip",suffix = "location", required = true)
@@ -64,7 +56,7 @@ public abstract class AbstractCompilationLocal {
     
     public String           gradleZipVersion;
     
-    String                  javaHomeCommand;
+    protected String        javaHomeCommand;
     
     /*
      * Initialise all resources needed for testing.
@@ -74,16 +66,19 @@ public abstract class AbstractCompilationLocal {
     public void setupTest() throws ResourceUnavailableException, IOException, LinuxManagerException, IpNetworkManagerException {
         javaHomeCommand = "export JAVA_HOME=" + getJavaInstallation().getJavaHome();
         
+        testRunDirectory = getLinuxImage().getHome().resolve(runName);
+        
+        setProjectDirectory();
+        
         gradleZipVersion = getGradleVersion();
-        simplatformParentDir = setupSimPlatform();
         gradleBin = installGradle();
     }
+    
     
     /*
      * Extracts the version from the gradle zip path
      */
-    
-    public String getGradleVersion() {
+    private String getGradleVersion() {
         final String regex = "gradle-(\\d+\\.\\d+\\.\\d+)-bin\\.zip";
 
         final Pattern pattern = Pattern.compile(regex, Pattern.MULTILINE);
@@ -95,48 +90,30 @@ public abstract class AbstractCompilationLocal {
     }
 
     /*
-     * Download simplatform (github main branch), unzipp simplatform, then structure 
-     * and edit the directory and files it in such a way that they are usable by gradle.
+     * Method for unpacking an archive on a remote image.
      * 
-     * @return  remoteDirectory     the parent directory of the structured, edited, and usable simplatform project.
-     * 
+     * @param	archive		Path to the remote archive
+     * @param	target		Path to the directory to which archive should be unpacked 
      */
-    private Path setupSimPlatform() throws ResourceUnavailableException, IOException, LinuxManagerException {
-        logger.info("Downloading simplatform repository archive - main branch.");
-        Path simplatformZip = downloadHttp("https://github.com/galasa-dev/simplatform/archive/main.zip");
-        logger.info("Unzipping simplatform repository archive");
-        Path simplatformDir = unpackArchiveToTemp(simplatformZip);
-        logger.info("Set up repository");
-        Path simplatformParent = structureSimplatform(simplatformDir);
+    protected void unpackOnRemote(Path archive, Path target) throws IpNetworkManagerException, LinuxManagerException {
+        logger.trace("Unzipping archive \"" + archive.toString() + "\" to \"" + target.toString() + "\"");
+        String unpackRC = getLinuxImage().getCommandShell().issueCommand(
+                "unzip -q " + archive.toString()
+                + " -d " + target.toString()
+                + "; echo RC=$?"
+        );
         
-        refactorSimplatform(simplatformParent);
-        
-        Path remoteDir = getLinuxImage().getHome().resolve(runName + "/simplatform");
-        
-        logger.info("Upload simplatform to remote image");
-        copyDirectory(simplatformParent, remoteDir);
-
-        return remoteDir;
+        logger.trace("Checking return code of unpack");
+        assertThat(unpackRC).isEqualToIgnoringWhitespace("RC=0");
     }
-    
-    /*
-     * Parent method for renaming of files and changing of prefixes.
-     * To be overridden by subclasses to make further changes (e.g. building from isolated zip requires more changes)  
-     * 
-     * @param   simplatformParent   the path to the parent directory of the simplatform project
-     */
-    protected void refactorSimplatform(Path simplatformParent) throws IOException {
-        renameFiles(simplatformParent);
-        changeAllPrefixes(simplatformParent);
-    }
-    
+        
     /*
      * Method for downloading a resource from a http location.
      * 
      * @param   downloadLocation    string containing the url of the resource to be downloaded
      * @return  tempFile            the temporary location of the file that was downloaded.
      */
-    private Path downloadHttp(String downloadLocation) throws ResourceUnavailableException {
+    protected Path downloadHttp(String downloadLocation) throws ResourceUnavailableException {
 
         logger.trace("Retrieving Http Resource: " + downloadLocation);
 
@@ -166,75 +143,30 @@ public abstract class AbstractCompilationLocal {
             throw new ResourceUnavailableException("Unable to download from: " + uri.toString(), e);
         }
     }
-    
-    /*
-     * Unpacks specified archive to a temporary location, then passes back the path to the temporary location.
-     * 
-     * @param   archiveFile             path to the archive to be downloaded
-     * @return  unpackedDir             path to the directory containing the unpacked archive
-     * @see     unpackSourceToTarget
-     */
-    private Path unpackArchiveToTemp(Path archiveFile) throws IOException {
-        logger.trace("Unzipping: " + archiveFile.toString());
-        Path unpackedDir = Files.createTempDirectory("galasa.test.compilation.unpacked");
-        unpackedDir.toFile().deleteOnExit();
         
-        unpackSourceToTarget(archiveFile, unpackedDir);
-        
-        return unpackedDir;
-        
-    }
-    
-    /*
-     * Unpacks the specified source archive to the specified target directory.
-     * 
-     * @param   source  the archive to be unpacked
-     * @param   target  the directory to unpack the source to
-     * 
-     */
-    private void unpackSourceToTarget(Path source, Path target) throws IOException {
-        logger.trace("Unpacking: " + source.toString() + " to " + target.toString());
-        ZipInputStream zipInputStream = new ZipInputStream(Files.newInputStream(source));
-        ZipEntry entry;
-        while ((entry = zipInputStream.getNextEntry()) != null) {
-            final Path entryTarget = target.resolve(entry.getName());
-            if (entry.isDirectory()) {
-                Files.createDirectory(entryTarget);
-            } else {
-                Files.copy(zipInputStream, entryTarget);
-            }
-        }
-    }
-    
     /*
      * Installs gradle (Download locally, upload to image, and unzip on image), 
      * then returns the path to the gradle bin directory. 
      * 
+     * @return	gradleWorkingDir	The path to the gradle installation directory.
      */
     private Path installGradle() throws ResourceUnavailableException, LinuxManagerException, IOException, IpNetworkManagerException {
         // Download Gradle
         logger.info("Installing Gradle");
         logger.trace("Downloading Gradle Zip from: " + gradleZipLocation);
-        Path gradleZip = downloadHttp(gradleZipLocation);
+        Path localGradleArchive = downloadHttp(gradleZipLocation);
         Path runLocation = getLinuxImage().getHome().resolve(runName);
-        Path remoteGradleZip = runLocation.resolve("gradle-" + gradleZipVersion + ".zip");
+        Path remoteGradleArchive = runLocation.resolve("gradle-" + gradleZipVersion + ".zip");
         Path remoteGradleDir = runLocation.resolve("gradle");
         
         // Upload Gradle
         logger.trace("Uploading gradle archive to remote image");
-        logger.trace("Copying: " + gradleZip.toString() + " to " + remoteGradleZip.toString());
-        Files.copy(gradleZip, remoteGradleZip);
+        logger.trace("Copying: " + localGradleArchive.toString() + " to " + remoteGradleArchive.toString());
+        Files.copy(localGradleArchive, remoteGradleArchive);
         
         // Unzip Gradle
         logger.trace("Unzipping gradle archive on remote image");
-        String unzipRC = getLinuxImage().getCommandShell().issueCommand(
-                "unzip -q " + remoteGradleZip.toString()
-                + " -d " + remoteGradleDir.toString()
-                + "; echo $?"
-        );
-        
-        logger.trace("Checking return code of unzip");
-        assertThat(unzipRC).isEqualToIgnoringWhitespace("0");
+        unpackOnRemote(remoteGradleArchive, remoteGradleDir);
         
         Path gradleWorkingDir = remoteGradleDir.resolve("gradle-" + gradleZipVersion + "/bin");
         
@@ -248,122 +180,10 @@ public abstract class AbstractCompilationLocal {
         
         return gradleWorkingDir;
     }
+        
     
     /*
-     * Retrieves the necessary source files from the unpacked simplatform archive and 
-     * structures them in a parent directory, along with a parent settings file.
-     * 
-     * @param   unpackedDir     The directory containing he unpackaged simplatform archive  
-     * @return  parentDir       The parent directory of the newly structured simplatform parent project
-     * 
-     */
-    private Path structureSimplatform(Path unpackedDir) throws IOException {
-        // Create new (temp) directory
-        logger.trace("Creating project parent directory (for manager and tests)");
-        Path parentDir = Files.createTempDirectory("galasa.test.simplatform.parent");
-        parentDir.toFile().deleteOnExit();
-        
-        // Create parent settings file
-        
-        logger.trace("Creating settings.gradle");
-        Path parentSettingsFile = parentDir.resolve("settings.gradle");
-        Files.createFile(parentSettingsFile);
-        StringBuilder settingsSB = new StringBuilder();
-        settingsSB.append("include '");
-        settingsSB.append(managerProjectName);
-        settingsSB.append("'\n");
-        settingsSB.append("include '");
-        settingsSB.append(testProjectName);
-        settingsSB.append("'\n");
-        Files.write(parentSettingsFile, settingsSB.toString().getBytes());
-        
-        // Get Manager Files
-        logger.trace("Copying managers source into parent directory");
-        copyDirectory(unpackedDir.resolve("simplatform-main/galasa-simbank-tests/" + managerProjectName + "/"), parentDir.resolve(managerProjectName));
-        // Get Tests
-        logger.trace("Copying tests source into parent directory");
-        copyDirectory(unpackedDir.resolve("simplatform-main/galasa-simbank-tests/" + testProjectName + "/"), parentDir.resolve(testProjectName));
-
-        return parentDir;
-    }
-    
-    /*
-     * Copy a directory to another directory using nio Files.
-     * 
-     * @param   source      The directory from which to copy files / directories  
-     * @param   target      The directory to which files / directories will be copied
-     * 
-     */
-    private void copyDirectory(Path source, Path target) throws IOException {
-        logger.info("Copying Directory: " + source.toString() + " to " + target.toString());        Files.walkFileTree(source, new SimpleFileVisitor<Path>() {
-            @Override
-            public FileVisitResult preVisitDirectory(Path dir, BasicFileAttributes attrs)
-                    throws IOException {
-                Files.createDirectories(target.resolve(source.relativize(dir).toString()));
-                return FileVisitResult.CONTINUE;
-            }
-
-            @Override
-            public FileVisitResult visitFile(Path file, BasicFileAttributes attrs)
-                    throws IOException {
-                Files.copy(file, target.resolve(source.relativize(file).toString()));
-                return FileVisitResult.CONTINUE;
-            }
-        });
-    }
-    
-    /*
-     * Ranames the example files with names that gradle will recognise.
-     * 
-     * @param   simplatformParent   The directory of the simplatform project
-     */
-    protected void renameFiles(Path simplatformParent) throws IOException {
-        logger.trace("Renaming example files");
-        Path managerDir = simplatformParent.resolve(managerProjectName);
-        Path testDir = simplatformParent.resolve(testProjectName);
-        
-        // Manager
-        Files.move(managerDir.resolve("settings-example.gradle"), managerDir.resolve("settings.gradle"));        
-        Files.move(managerDir.resolve("build-example.gradle"), managerDir.resolve("build.gradle"));
-        Files.move(managerDir.resolve("bnd-example.bnd"), managerDir.resolve("bnd.bnd"));
-        
-        // Tests
-        Files.move(testDir.resolve("settings-example.gradle"), testDir.resolve("settings.gradle"));        
-        Files.move(testDir.resolve("build-example.gradle"), testDir.resolve("build.gradle"));
-        Files.move(testDir.resolve("bnd-example.bnd"), testDir.resolve("bnd.bnd"));
-        
-    }
-    
-    /*
-     * Ranames the prefix placeholders within the build and settings files of the 
-     * manager and test projects.
-     * 
-     * @param   simplatformParent   The directory of the simplatform project
-     */
-    protected void changeAllPrefixes(Path simplatformParent) throws IOException {
-        // Manager
-        changePrefix(simplatformParent.resolve(managerProjectName + "/build.gradle"));
-        changePrefix(simplatformParent.resolve(managerProjectName + "/settings.gradle"));
-        // Tests
-        changePrefix(simplatformParent.resolve(testProjectName + "/build.gradle"));
-        changePrefix(simplatformParent.resolve(testProjectName + "/settings.gradle"));
-    }
-    
-    /*
-     * Ranames the prefix placeholders within the specified file
-     * 
-     * @param   file   The file in which the prefix will be changed.
-     */
-    private void changePrefix(Path file) throws IOException {
-        String incumbent = "%%prefix%%";
-        String fileData = new String(Files.readAllBytes(file), Charset.defaultCharset());
-        fileData = fileData.replace(incumbent, prefix);
-        Files.write(file, fileData.getBytes());
-        logger.trace("Changing prefix (" + incumbent + ") to \"" + prefix + "\" in file: " + file.toString());
-    }
-    
-    /*
-     * Runs and gradle build against the prepared simplatform code.
+     * Runs and Gradle build against the prepared Simplatform code.
      * Passes if "BUILD SUCCESSFUL" appears in the output.
      * 
      */
@@ -373,17 +193,19 @@ public abstract class AbstractCompilationLocal {
         
         logger.info("Running Gradle Build");
         
+        // Set Java Home, go to project directory, execute the unpackaged Gradle binary.
+        // Pass to the Gradle binary: 
+        // 		* User home directory
+        // 		* Option to ensure output is logger friendly
+        //		* The task(s) to be executed.
         String buildCommand = javaHomeCommand + "; "
-                // Go to project directory
-                + "cd " + simplatformParentDir.toString() + "; "
-                // Build using Gradle
+                + "cd " + projectDirectory.toString() + "; "
                 + gradleBin.toString() + "/gradle "
-                + "-Dgradle.user.home=" + getLinuxImage().getHome().resolve(runName) + "/.gradle"
+                + "-Dgradle.user.home=" + testRunDirectory + "/.gradle"
                 + "--console "
                 + "build";
         
-        logger.info("Issuing Command:");
-        logger.info(buildCommand);
+        logger.info("Issuing Command: " + buildCommand);
         
         String managerBuildResults = getLinuxImage().getCommandShell().issueCommand(buildCommand);
         
@@ -391,19 +213,26 @@ public abstract class AbstractCompilationLocal {
     }
     
     /*
-     * @return  ecosystem   the ecosystem instance associated with the test.
+     * Sets the project directory to run the gradle build against
+     * 
+     */
+    abstract protected void setProjectDirectory() throws ResourceUnavailableException, LinuxManagerException, IpNetworkManagerException, IOException;
+
+    
+    /*
+     * @return  ecosystem	The ecosystem instance associated with the test.
      * 
      */
     abstract protected IGenericEcosystem getEcosystem();
     
     /*
-     * @return  linuxImage  the linux image instance associated with the test.
+     * @return  linuxImage  The linux image instance associated with the test.
      * 
      */
     abstract protected ILinuxImage getLinuxImage();
 
     /*
-     * @return  javaInstallation    the java installation instance associated with the test.
+     * @return  javaInstallation    The java installation instance associated with the test.
      * 
      */
     abstract protected IJavaUbuntuInstallation getJavaInstallation();
