+++ galasa-inttests-parent/dev.galasa.inttests/src/main/java/dev/galasa/inttests/compilation/simbank/local/CompilationLocalJava08Ubuntu.java
@@ -1,0 +1,49 @@
+/*
+ * Licensed Materials - Property of IBM
+ * 
+ * (c) Copyright IBM Corp. 2021.
+ */
+package dev.galasa.inttests.compilation.simbank.local;
+
+import dev.galasa.Test;
+import dev.galasa.TestAreas;
+import dev.galasa.galasaecosystem.IGenericEcosystem;
+import dev.galasa.galasaecosystem.ILocalEcosystem;
+import dev.galasa.galasaecosystem.LocalEcosystem;
+import dev.galasa.inttests.compilation.simbank.AbstractCompilationLocalSimBank;
+import dev.galasa.java.JavaVersion;
+import dev.galasa.java.ubuntu.IJavaUbuntuInstallation;
+import dev.galasa.java.ubuntu.JavaUbuntuInstallation;
+import dev.galasa.linux.ILinuxImage;
+import dev.galasa.linux.LinuxImage;
+import dev.galasa.linux.OperatingSystem;
+
+@Test
+@TestAreas({"compilation","localecosystem","java08","ubuntu"})
+public class CompilationLocalJava08Ubuntu extends AbstractCompilationLocalSimBank {
+
+    @LocalEcosystem(linuxImageTag = "PRIMARY")
+    public ILocalEcosystem ecosystem;
+    
+    @LinuxImage(operatingSystem = OperatingSystem.ubuntu)
+    public ILinuxImage linuxImage;
+    
+    @JavaUbuntuInstallation(javaVersion = JavaVersion.v8)
+    public IJavaUbuntuInstallation java;
+
+    @Override
+    protected IGenericEcosystem getEcosystem() {
+        return this.ecosystem;
+    }
+
+    @Override
+    protected ILinuxImage getLinuxImage() {
+        return this.linuxImage;
+    }
+    
+    @Override
+    protected IJavaUbuntuInstallation getJavaInstallation() {
+        return this.java;
+    }
+
+}
