+++ galasa-inttests-parent/dev.galasa.inttests/src/main/java/dev/galasa/inttests/compilation/simbank/local/CompilationLocalJava11Ubuntu.java
@@ -3,14 +3,14 @@
  * 
  * (c) Copyright IBM Corp. 2021.
  */
-package dev.galasa.inttests.compilation.local;
+package dev.galasa.inttests.compilation.simbank.local;
 
 import dev.galasa.Test;
 import dev.galasa.TestAreas;
 import dev.galasa.galasaecosystem.IGenericEcosystem;
 import dev.galasa.galasaecosystem.ILocalEcosystem;
 import dev.galasa.galasaecosystem.LocalEcosystem;
-import dev.galasa.inttests.compilation.AbstractCompilationLocal;
+import dev.galasa.inttests.compilation.simbank.AbstractCompilationLocalSimBank;
 import dev.galasa.java.JavaVersion;
 import dev.galasa.java.ubuntu.IJavaUbuntuInstallation;
 import dev.galasa.java.ubuntu.JavaUbuntuInstallation;
@@ -20,7 +20,7 @@ import dev.galasa.linux.OperatingSystem;
 
 @Test
 @TestAreas({"compilation","localecosystem","java11","ubuntu"})
-public class CompilationLocalJava11Ubuntu extends AbstractCompilationLocal {
+public class CompilationLocalJava11Ubuntu extends AbstractCompilationLocalSimBank {
 
     @LocalEcosystem(linuxImageTag = "PRIMARY")
     public ILocalEcosystem ecosystem;
