+++ galasa-inttests-parent/dev.galasa.inttests/src/main/java/dev/galasa/inttests/compilation/simbank/local/isolated/CompilationLocalJava08UbuntuIsolated.java
@@ -3,12 +3,7 @@
  * 
  * (c) Copyright IBM Corp. 2021.
  */
-package dev.galasa.inttests.compilation.local.isolated;
-
-import java.io.IOException;
-import java.nio.charset.Charset;
-import java.nio.file.Files;
-import java.nio.file.Path;
+package dev.galasa.inttests.compilation.simbank.local.isolated;
 
 import dev.galasa.Test;
 import dev.galasa.TestAreas;
@@ -16,7 +11,7 @@ import dev.galasa.galasaecosystem.IGenericEcosystem;
 import dev.galasa.galasaecosystem.ILocalEcosystem;
 import dev.galasa.galasaecosystem.IsolationInstallation;
 import dev.galasa.galasaecosystem.LocalEcosystem;
-import dev.galasa.inttests.compilation.AbstractCompilationLocalZip;
+import dev.galasa.inttests.compilation.simbank.AbstractCompilationLocalSimBankOffline;
 import dev.galasa.java.JavaVersion;
 import dev.galasa.java.ubuntu.IJavaUbuntuInstallation;
 import dev.galasa.java.ubuntu.JavaUbuntuInstallation;
@@ -26,7 +21,7 @@ import dev.galasa.linux.OperatingSystem;
 
 @Test
 @TestAreas({"compilation","localecosystem","java08","ubuntu","isolated"})
-public class CompilationLocalJava08UbuntuIsolated extends AbstractCompilationLocalZip {
+public class CompilationLocalJava08UbuntuIsolated extends AbstractCompilationLocalSimBankOffline {
 
     @LocalEcosystem(linuxImageTag = "PRIMARY", isolationInstallation = IsolationInstallation.Full)
     public ILocalEcosystem ecosystem;
@@ -36,31 +31,6 @@ public class CompilationLocalJava08UbuntuIsolated extends AbstractCompilationLoc
     
     @JavaUbuntuInstallation(javaVersion = JavaVersion.v8)
     public IJavaUbuntuInstallation java;
-
-    @Override
-    protected void refactorSimplatform(Path simplatformParent) throws IOException {
-        renameFiles(simplatformParent);
-        changeAllPrefixes(simplatformParent);
-        
-        Path managerBuildGradle = simplatformParent.resolve("dev.galasa.simbank.manager/build.gradle");
-        Path testBuildGradle = simplatformParent.resolve("dev.galasa.simbank.tests/build.gradle");
-        Path parentSettings = simplatformParent.resolve("settings.gradle");
-        
-        // Alter project parent
-        addPluginManagementRepo(parentSettings);
-        
-        // Alter manager project
-        updateMavenRepo(managerBuildGradle); 
-        addDependencyConstraints(managerBuildGradle);
-        
-        // Alter test project
-        updateMavenRepo(testBuildGradle);
-        // Add a list of managers to the test(s)
-        addManagerDependencies(testBuildGradle, allManagers);
-        addDependencyConstraints(testBuildGradle);
-        addImplementationConstraints(testBuildGradle);
-        
-    }
     
     @Override
     protected IGenericEcosystem getEcosystem() {
