+++ galasa-inttests-parent/dev.galasa.inttests/src/main/java/dev/galasa/inttests/compilation/simbank/local/mvp/CompilationLocalJava11UbuntuMvp.java
@@ -3,14 +3,15 @@
  * 
  * (c) Copyright IBM Corp. 2021.
  */
-package dev.galasa.inttests.compilation.local;
+package dev.galasa.inttests.compilation.simbank.local.mvp;
 
 import dev.galasa.Test;
 import dev.galasa.TestAreas;
 import dev.galasa.galasaecosystem.IGenericEcosystem;
 import dev.galasa.galasaecosystem.ILocalEcosystem;
+import dev.galasa.galasaecosystem.IsolationInstallation;
 import dev.galasa.galasaecosystem.LocalEcosystem;
-import dev.galasa.inttests.compilation.AbstractCompilationLocal;
+import dev.galasa.inttests.compilation.simbank.AbstractCompilationLocalSimBankOffline;
 import dev.galasa.java.JavaVersion;
 import dev.galasa.java.ubuntu.IJavaUbuntuInstallation;
 import dev.galasa.java.ubuntu.JavaUbuntuInstallation;
@@ -19,18 +20,18 @@ import dev.galasa.linux.LinuxImage;
 import dev.galasa.linux.OperatingSystem;
 
 @Test
-@TestAreas({"compilation","localecosystem","java08","ubuntu"})
-public class CompilationLocalJava08Ubuntu extends AbstractCompilationLocal {
+@TestAreas({"compilation","localecosystem","java11","ubuntu","mvp"})
+public class CompilationLocalJava11UbuntuMvp extends AbstractCompilationLocalSimBankOffline {
 
-    @LocalEcosystem(linuxImageTag = "PRIMARY")
+    @LocalEcosystem(linuxImageTag = "PRIMARY", isolationInstallation = IsolationInstallation.Mvp)
     public ILocalEcosystem ecosystem;
     
-    @LinuxImage(operatingSystem = OperatingSystem.ubuntu)
+    @LinuxImage(operatingSystem = OperatingSystem.ubuntu, capabilities = "isolated")
     public ILinuxImage linuxImage;
     
-    @JavaUbuntuInstallation(javaVersion = JavaVersion.v8)
+    @JavaUbuntuInstallation(javaVersion = JavaVersion.v11)
     public IJavaUbuntuInstallation java;
-
+    
     @Override
     protected IGenericEcosystem getEcosystem() {
         return this.ecosystem;
